update report api

+++ build/src/controller/report.controller.js
@@ -2,11 +2,10 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.ReportController = void 0;
 const lib_1 = require("../lib");
+const io_1 = require("../lib/socket/io");
 const schema_1 = require("../schema");
 const service_1 = require("../service");
 const utils_1 = require("../utils");
-const message_controller_1 = require("./message.controller");
-const io_1 = require("../lib/socket/io");
 class ReportController {
     constructor() {
         this.getAllReports = async (req, res) => {
@@ -50,44 +49,61 @@ class ReportController {
             const report = await this.reportService.updateReport(id, parsed.data);
             return (0, utils_1.SuccessResponse)({ res, data: report, statusCode: 201 });
         };
-        this.updateStatusChatOwner = async (req, res) => {
+        this.updateStatusChatOwner = async (req, res, next) => {
             const id = Number(req.params.id);
+            console.log("Params i", id);
             const parsed = schema_1.updateStatusReportSchema.safeParse(req.body);
+            const reportExist = await this.reportService.findReportById(id);
+            if (!reportExist) {
+                next("Report does not exist with that id");
+            }
             // ** Add logic to send notification to the owner that own that report
             // ** User Id here is the one that report
             // ** Send Notification to the owner report that someone has found your
             if (parsed.data && parsed.success) {
-                const report = await this.reportService.updateReportStatus(id, parseInt(parsed.data.userId), "CHATOWNER");
+                const report = await this.reportService.updateReportStatus(id, parsed.data.userId, "CHATOWNER");
                 const userIdreport = report.userId;
                 console.log(">>> Send notification to ", userIdreport);
                 // Create empty chat (optional: after confirming, notify/report owner)
-                const senderId = parseInt(parsed.data.userId); // user who confirms
+                const senderId = parsed.data.userId; // user who confirms
                 const receiverId = userIdreport; // owner of the report
                 await this.createChatBetweenUser(senderId, receiverId);
                 return (0, utils_1.SuccessResponse)({ res, data: report, statusCode: 201 });
             }
         };
-        this.updateStatusConfirm = async (req, res) => {
+        this.updateStatusConfirm = async (req, res, next) => {
             const id = Number(req.params.id);
             const parsed = schema_1.updateStatusReportSchema.safeParse(req.body);
+            const reportExist = await this.reportService.findReportById(id);
+            if (!reportExist) {
+                next("Report does not exist with that id");
+            }
             // ** User Id here is the one that confirm
             // ** Send Notification to the user that you success earn a badge
             if (parsed.data && parsed.success) {
-                const report = await this.reportService.updateReportStatus(id, parseInt(parsed.data.userId), "COMPLETED");
+                const report = await this.reportService.updateReportStatus(id, parsed.data.userId, "COMPLETED");
                 return (0, utils_1.SuccessResponse)({ res, data: report, statusCode: 201 });
             }
         };
-        this.updateStatusCancel = async (req, res) => {
+        this.updateStatusCancel = async (req, res, next) => {
             const id = Number(req.params.id);
             const parsed = schema_1.updateStatusReportSchema.safeParse(req.body);
+            const reportExist = await this.reportService.findReportById(id);
+            if (!reportExist) {
+                next("Report does not exist with that id");
+            }
             // ** Need user id just incase not delete that report
             if (parsed.data && parsed.success) {
                 const report = await this.reportService.deleteReport(id);
                 return (0, utils_1.SuccessResponse)({ res, data: report, statusCode: 201 });
             }
         };
-        this.deleteReport = async (req, res) => {
+        this.deleteReport = async (req, res, next) => {
             const id = Number(req.params.id);
+            const reportExist = await this.reportService.findReportById(id);
+            if (!reportExist) {
+                next("Report does not exist with that id");
+            }
             const report = await this.reportService.deleteReport(id);
             return (0, utils_1.SuccessResponse)({
                 res,
@@ -98,7 +114,6 @@ class ReportController {
         this.reportService = new service_1.ReportService(lib_1.PrismaClient);
         this.userService = new service_1.UserService(lib_1.PrismaClient);
         this.messageService = new service_1.MessageService(lib_1.PrismaClient);
-        this.messageController = new message_controller_1.MessageController();
     }
     async createChatBetweenUser(senderId, receiverId) {
         try {
